Resume entity extraction/requirements.txt
docx2txt==0.8
pdfminer.six==20211012
nltk==3.6.3
flask==0.10.1
pandas
numpy
transformers==4.12.5
flair==0.10
python-dateutil==2.8.2
